Refactor : getDiffTime -> time 파일 이름 변경 - 시간 관련 함수를 한 파일에 모아 관리하기 위해 파일명을 보다 명확하게 변경함
