+++ Client/src/pages/issue/IssueDetailPage.jsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import IssueCommentForm from '../../components/issue/IssueCommentForm';
 import { GET_ISSUE } from '../../utils/api.js';
 import { getOptions } from '../../utils/fetchOptions';
-import getDiffTime from '../../utils/getDiffTime';
+import { getDiffTime } from '../../utils/time';
 import Container from '../../components/shared/container/Container';
 import IssueStateButton from '../../components/issue/IssueStateButton';
